+++ packages/slate-plugins/src/index.tsx
@@ -4,6 +4,7 @@
 
 export * from '@udecode/slate-plugins-alignment';
 export * from '@udecode/slate-plugins-alignment-ui';
+export * from '@udecode/slate-plugins-ast-serializer';
 export * from '@udecode/slate-plugins-autoformat';
 export * from '@udecode/slate-plugins-basic-elements';
 export * from '@udecode/slate-plugins-basic-marks';