+++ packages/elements/list-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @udecode/slate-plugins-list-ui
 
+## 1.0.0-next.48
+
+### Patch Changes
+
+- Updated dependencies [[`a15ab621`](https://github.com/udecode/slate-plugins/commit/a15ab6217c6e2d4eb2a1320f6b76c483fc963047)]:
+  - @udecode/slate-plugins-list@1.0.0-next.48
+
+## 1.0.0-next.46
+
+### Patch Changes
+
+- Updated dependencies [[`6e9068f6`](https://github.com/udecode/slate-plugins/commit/6e9068f6f483b698b6b3aae6819333103504f41b)]:
+  - @udecode/slate-plugins-common@1.0.0-next.46
+  - @udecode/slate-plugins-list@1.0.0-next.46
+  - @udecode/slate-plugins-toolbar@1.0.0-next.46
+
 ## 1.0.0-next.43
 
 ### Patch Changes