+++ packages/reset-node/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @udecode/slate-plugins-reset-node
 
+## 1.0.0-next.50
+
+### Minor Changes
+
+- [#824](https://github.com/udecode/slate-plugins/pull/824) [`92e19158`](https://github.com/udecode/slate-plugins/commit/92e19158fe6edf93c238e5de9727505967071b96) Thanks [@stephenkiers](https://github.com/stephenkiers)! - Allow `li` elements to be deleted, even when they are the first node
+
+## 1.0.0-next.46
+
+### Patch Changes
+
+- Updated dependencies [[`6e9068f6`](https://github.com/udecode/slate-plugins/commit/6e9068f6f483b698b6b3aae6819333103504f41b)]:
+  - @udecode/slate-plugins-common@1.0.0-next.46
+
 ## 1.0.0-next.40
 
 ### Patch Changes