+++ packages/serializers/md-serializer/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @udecode/slate-plugins-md-serializer
 
+## 1.0.0-next.48
+
+### Minor Changes
+
+- [#821](https://github.com/udecode/slate-plugins/pull/821) [`091f0940`](https://github.com/udecode/slate-plugins/commit/091f0940bd3c06c3dfaf49a4ab14eb611678637d) Thanks [@dylans](https://github.com/dylans)! - If empty fragment, eject from deserializer
+
+### Patch Changes
+
+- [#819](https://github.com/udecode/slate-plugins/pull/819) [`b82f47a6`](https://github.com/udecode/slate-plugins/commit/b82f47a66ea1521dc426ae87e1ec37f004593cbe) Thanks [@dylans](https://github.com/dylans)! - Markdown deserializer was breaking the pasting of a simple URL into the editor. Now checks the content and if it's simply a URL, it skips the handling of the content as markdown
+
+- Updated dependencies [[`a15ab621`](https://github.com/udecode/slate-plugins/commit/a15ab6217c6e2d4eb2a1320f6b76c483fc963047)]:
+  - @udecode/slate-plugins-list@1.0.0-next.48
+
+## 1.0.0-next.47
+
+### Minor Changes
+
+- [#813](https://github.com/udecode/slate-plugins/pull/813) [`2d671565`](https://github.com/udecode/slate-plugins/commit/2d67156509ca8689aede2d4a9a45f749772c789c) Thanks [@dylans](https://github.com/dylans)! - Fix ast-deserialize insert, minor cleanup to html/md deserializer
+
+- [#814](https://github.com/udecode/slate-plugins/pull/814) [`1bbdae38`](https://github.com/udecode/slate-plugins/commit/1bbdae389e7ec3ec7a54877526055a9464e46fdf) Thanks [@dylans](https://github.com/dylans)! - allow override of fragment root for deserializers
+
+## 1.0.0-next.46
+
+### Patch Changes
+
+- Updated dependencies [[`6e9068f6`](https://github.com/udecode/slate-plugins/commit/6e9068f6f483b698b6b3aae6819333103504f41b)]:
+  - @udecode/slate-plugins-common@1.0.0-next.46
+  - @udecode/slate-plugins-block-quote@1.0.0-next.46
+  - @udecode/slate-plugins-code-block@1.0.0-next.46
+  - @udecode/slate-plugins-heading@1.0.0-next.46
+  - @udecode/slate-plugins-link@1.0.0-next.46
+  - @udecode/slate-plugins-list@1.0.0-next.46
+  - @udecode/slate-plugins-paragraph@1.0.0-next.46
+
+## 1.0.0-next.44
+
+### Patch Changes
+
+- [#801](https://github.com/udecode/slate-plugins/pull/801) [`7f5f223d`](https://github.com/udecode/slate-plugins/commit/7f5f223d39e0b6a6381d42d1a95d73592960319a) Thanks [@dylans](https://github.com/dylans)! - Make markdown deserializer more consistent in approach with html and ast deserializers
+
 ## 1.0.0-next.43
 
 ### Patch Changes